pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gfpipeline"
version = "0.1.0"
description = "Gene Family Analysis Pipeline CLI"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "biopython>=1.80",
    "requests>=2.28",
    "matplotlib>=3.6",
    "seaborn>=0.12",
    "matplotlib-venn>=0.11",
    "hypothesis>=6.0",
]

[project.scripts]
gfpipeline = "gfpipeline.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["gfpipeline*"]
